requirements.txt
PyOpenSSL

sympy==1.3
jupyter==1.0.0

TA-Lib == 0.4.17
psycopg2-binary==2.7.5
SQLAlchemy==1.2.10

pymongo==3.7.1

numpy
pandas
scipy
matplotlib
seaborn
# django 模块
django==2.1.2
django-crontab==0.7.1

mpl_finance==0.10.0
pandas_datareader==0.7.0
tushare==1.2.12

statsmodels==0.9.0

PyQt5==5.11.2

python-dateutil==2.7.3
Logbook==1.4.0

#abupy depend
ipython==6.5.0
ipywidgets==7.4.2
bokeh==0.13.0
toolz==0.9.0
Pillow==5.2.0
numba==0.40.0
scikit-image==0.14.0
scikit-learn==0.19.2
six==1.11.0
psutil==5.4.7

pytest
pytdx==1.67
easyquotation==0.5.12
PyAlgoTrade==0.20


pypinyin
cython
pyecharts_jupyter_installer


jieba
fuzzyfinder

xlrd # excel support
jqdatasdk


tornado==5.1.1
influxdb==5.2.0






